requirements.txt
django
ipython
django-colorfield